@@ -56,7 +56,7 @@
     "@storybook/angular": "7.6.19",
     "@storybook/jest": "0.2.3",
     "@storybook/testing-library": "0.2.2",
-    "@types/argon2-browser": "1.18.1",
+    "@types/argon2-browser": "1.18.4",
     "@types/chrome": "0.0.262",
     "@types/firefox-webext-browser": "111.0.5",
     "@types/inquirer": "8.2.10",
@@ -82,7 +82,7 @@
     "@typescript-eslint/eslint-plugin": "7.13.1",
     "@typescript-eslint/parser": "7.13.1",
     "@webcomponents/custom-elements": "1.6.0",
-    "@yao-pkg/pkg": "^5.11.5",
+    "@yao-pkg/pkg": "^5.12.0",
     "autoprefixer": "10.4.19",
     "babel-loader": "^9.1.3",
     "base64-loader": "1.0.0",
@@ -92,7 +92,7 @@
     "copy-webpack-plugin": "12.0.2",
     "cross-env": "7.0.3",
     "css-loader": "6.10.0",
-    "electron": "30.1.2",
+    "electron": "31.1.0",
     "electron-builder": "24.13.3",
     "electron-log": "5.0.1",
     "electron-reload": "2.0.0-alpha.1",
