@@ -22,7 +22,7 @@ import { UnencryptedMessageResponse } from "../models/native-messaging/unencrypt
 
 import { EncryptedMessageHandlerService } from "./encrypted-message-handler.service";
 
-const EncryptionAlgorithm = "sha1";
+const HashAlgorithmForAsymmetricEncryption = "sha1";
 
 // This service handles messages using the protocol created for the DuckDuckGo integration.
 @Injectable()
@@ -117,7 +117,7 @@ export class NativeMessageHandlerService {
       const encryptedSecret = await this.cryptoFunctionService.rsaEncrypt(
         secret,
         remotePublicKey,
-        EncryptionAlgorithm,
+        HashAlgorithmForAsymmetricEncryption,
       );
 
       this.sendResponse({
