@@ -3,7 +3,7 @@ import { RouterModule, Routes } from "@angular/router";
 
 import { canAccessMembersTab } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 
-import { OrganizationPermissionsGuard } from "../guards/org-permissions.guard";
+import { organizationPermissionsGuard } from "../guards/org-permissions.guard";
 
 import { MembersComponent } from "./members.component";
 
@@ -11,10 +11,9 @@ const routes: Routes = [
   {
     path: "",
     component: MembersComponent,
-    canActivate: [OrganizationPermissionsGuard],
+    canActivate: [organizationPermissionsGuard(canAccessMembersTab)],
     data: {
       titleId: "members",
-      organizationPermissions: canAccessMembersTab,
     },
   },
 ];
