@@ -2,6 +2,7 @@ import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
 import { AuthGuard } from "@bitwarden/angular/auth/guards";
+import { featureFlaggedRoute } from "@bitwarden/angular/platform/utils/feature-flagged-route";
 import {
   canAccessOrgAdmin,
   canAccessGroupsTab,
@@ -11,11 +12,13 @@ import {
   canAccessSettingsTab,
 } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
+import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
 
-import { OrganizationPermissionsGuard } from "../../admin-console/organizations/guards/org-permissions.guard";
-import { OrganizationRedirectGuard } from "../../admin-console/organizations/guards/org-redirect.guard";
+import { organizationPermissionsGuard } from "../../admin-console/organizations/guards/org-permissions.guard";
+import { organizationRedirectGuard } from "../../admin-console/organizations/guards/org-redirect.guard";
 import { OrganizationLayoutComponent } from "../../admin-console/organizations/layouts/organization-layout.component";
 import { GroupsComponent } from "../../admin-console/organizations/manage/groups.component";
+import { NewGroupsComponent } from "../../admin-console/organizations/manage/new-groups.component";
 import { deepLinkGuard } from "../../auth/guards/deep-link.guard";
 import { VaultModule } from "../../vault/org-vault/vault.module";
 
@@ -23,18 +26,12 @@ const routes: Routes = [
   {
     path: ":organizationId",
     component: OrganizationLayoutComponent,
-    canActivate: [deepLinkGuard(), AuthGuard, OrganizationPermissionsGuard],
-    data: {
-      organizationPermissions: canAccessOrgAdmin,
-    },
+    canActivate: [deepLinkGuard(), AuthGuard, organizationPermissionsGuard(canAccessOrgAdmin)],
     children: [
       {
         path: "",
         pathMatch: "full",
-        canActivate: [OrganizationRedirectGuard],
-        data: {
-          autoRedirectCallback: getOrganizationRoute,
-        },
+        canActivate: [organizationRedirectGuard(getOrganizationRoute)],
         children: [], // This is required to make the auto redirect work, },
       },
       {
@@ -52,15 +49,18 @@ const routes: Routes = [
         path: "members",
         loadChildren: () => import("./members").then((m) => m.MembersModule),
       },
-      {
-        path: "groups",
-        component: GroupsComponent,
-        canActivate: [OrganizationPermissionsGuard],
-        data: {
-          titleId: "groups",
-          organizationPermissions: canAccessGroupsTab,
+      ...featureFlaggedRoute({
+        defaultComponent: GroupsComponent,
+        flaggedComponent: NewGroupsComponent,
+        featureFlag: FeatureFlag.GroupsComponentRefactor,
+        routeOptions: {
+          path: "groups",
+          canActivate: [organizationPermissionsGuard(canAccessGroupsTab)],
+          data: {
+            titleId: "groups",
+          },
         },
-      },
+      }),
       {
         path: "reporting",
         loadChildren: () =>
