@@ -1,3 +1,4 @@
+import { ScrollingModule } from "@angular/cdk/scrolling";
 import { NgModule } from "@angular/core";
 
 import { LooseComponentsModule } from "../../shared";
@@ -5,6 +6,7 @@ import { LooseComponentsModule } from "../../shared";
 import { CoreOrganizationModule } from "./core";
 import { GroupAddEditComponent } from "./manage/group-add-edit.component";
 import { GroupsComponent } from "./manage/groups.component";
+import { NewGroupsComponent } from "./manage/new-groups.component";
 import { OrganizationsRoutingModule } from "./organization-routing.module";
 import { SharedOrganizationModule } from "./shared";
 import { AccessSelectorModule } from "./shared/components/access-selector";
@@ -16,7 +18,8 @@ import { AccessSelectorModule } from "./shared/components/access-selector";
     CoreOrganizationModule,
     OrganizationsRoutingModule,
     LooseComponentsModule,
+    ScrollingModule,
   ],
-  declarations: [GroupsComponent, GroupAddEditComponent],
+  declarations: [GroupsComponent, NewGroupsComponent, GroupAddEditComponent],
 })
 export class OrganizationModule {}
