@@ -2,9 +2,8 @@ import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
 import { canAccessBillingTab } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
-import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
 
-import { OrganizationPermissionsGuard } from "../../admin-console/organizations/guards/org-permissions.guard";
+import { organizationPermissionsGuard } from "../../admin-console/organizations/guards/org-permissions.guard";
 import { organizationIsUnmanaged } from "../../billing/guards/organization-is-unmanaged.guard";
 import { WebPlatformUtilsService } from "../../core/web-platform-utils.service";
 import { PaymentMethodComponent } from "../shared";
@@ -16,8 +15,7 @@ import { OrganizationSubscriptionSelfhostComponent } from "./organization-subscr
 const routes: Routes = [
   {
     path: "",
-    canActivate: [OrganizationPermissionsGuard],
-    data: { organizationPermissions: canAccessBillingTab },
+    canActivate: [organizationPermissionsGuard(canAccessBillingTab)],
     children: [
       { path: "", pathMatch: "full", redirectTo: "subscription" },
       {
@@ -30,19 +28,23 @@ const routes: Routes = [
       {
         path: "payment-method",
         component: PaymentMethodComponent,
-        canActivate: [OrganizationPermissionsGuard, organizationIsUnmanaged],
+        canActivate: [
+          organizationPermissionsGuard((org) => org.canEditPaymentMethods),
+          organizationIsUnmanaged,
+        ],
         data: {
           titleId: "paymentMethod",
-          organizationPermissions: (org: Organization) => org.canEditPaymentMethods,
         },
       },
       {
         path: "history",
         component: OrgBillingHistoryViewComponent,
-        canActivate: [OrganizationPermissionsGuard, organizationIsUnmanaged],
+        canActivate: [
+          organizationPermissionsGuard((org) => org.canViewBillingHistory),
+          organizationIsUnmanaged,
+        ],
         data: {
           titleId: "billingHistory",
-          organizationPermissions: (org: Organization) => org.canViewBillingHistory,
         },
       },
     ],
