@@ -3,8 +3,7 @@ import { RouterModule, Routes } from "@angular/router";
 
 import { AuthGuard } from "@bitwarden/angular/auth/guards";
 import { canAccessSettingsTab } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
-import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
-import { OrganizationPermissionsGuard } from "@bitwarden/web-vault/app/admin-console/organizations/guards/org-permissions.guard";
+import { organizationPermissionsGuard } from "@bitwarden/web-vault/app/admin-console/organizations/guards/org-permissions.guard";
 import { OrganizationLayoutComponent } from "@bitwarden/web-vault/app/admin-console/organizations/layouts/organization-layout.component";
 
 import { SsoComponent } from "../../auth/sso/sso.component";
@@ -16,40 +15,34 @@ const routes: Routes = [
   {
     path: "organizations/:organizationId",
     component: OrganizationLayoutComponent,
-    canActivate: [AuthGuard, OrganizationPermissionsGuard],
+    canActivate: [AuthGuard, organizationPermissionsGuard()],
     children: [
       {
         path: "settings",
-        canActivate: [OrganizationPermissionsGuard],
-        data: {
-          organizationPermissions: canAccessSettingsTab,
-        },
+        canActivate: [organizationPermissionsGuard(canAccessSettingsTab)],
         children: [
           {
             path: "domain-verification",
             component: DomainVerificationComponent,
-            canActivate: [OrganizationPermissionsGuard],
+            canActivate: [organizationPermissionsGuard((org) => org.canManageDomainVerification)],
             data: {
               titleId: "domainVerification",
-              organizationPermissions: (org: Organization) => org.canManageDomainVerification,
             },
           },
           {
             path: "sso",
             component: SsoComponent,
-            canActivate: [OrganizationPermissionsGuard],
+            canActivate: [organizationPermissionsGuard((org) => org.canManageSso)],
             data: {
               titleId: "singleSignOn",
-              organizationPermissions: (org: Organization) => org.canManageSso,
             },
           },
           {
             path: "scim",
             component: ScimComponent,
-            canActivate: [OrganizationPermissionsGuard],
+            canActivate: [organizationPermissionsGuard((org) => org.canManageScim)],
             data: {
               titleId: "scim",
-              organizationPermissions: (org: Organization) => org.canManageScim,
             },
           },
           {
@@ -58,9 +51,8 @@ const routes: Routes = [
               import("./manage/device-approvals/device-approvals.component").then(
                 (mod) => mod.DeviceApprovalsComponent,
               ),
-            canActivate: [OrganizationPermissionsGuard],
+            canActivate: [organizationPermissionsGuard((org) => org.canManageDeviceApprovals)],
             data: {
-              organizationPermissions: (org: Organization) => org.canManageDeviceApprovals,
               titleId: "deviceApprovals",
             },
           },
