@@ -1,8 +1,7 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
-import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
-import { OrganizationPermissionsGuard } from "@bitwarden/web-vault/app/admin-console/organizations/guards/org-permissions.guard";
+import { organizationPermissionsGuard } from "@bitwarden/web-vault/app/admin-console/organizations/guards/org-permissions.guard";
 
 import { SecretsManagerExportComponent } from "./porting/sm-export.component";
 import { SecretsManagerImportComponent } from "./porting/sm-import.component";
@@ -11,19 +10,17 @@ const routes: Routes = [
   {
     path: "import",
     component: SecretsManagerImportComponent,
-    canActivate: [OrganizationPermissionsGuard],
+    canActivate: [organizationPermissionsGuard((org) => org.isAdmin)],
     data: {
       titleId: "importData",
-      organizationPermissions: (org: Organization) => org.isAdmin,
     },
   },
   {
     path: "export",
     component: SecretsManagerExportComponent,
-    canActivate: [OrganizationPermissionsGuard],
+    canActivate: [organizationPermissionsGuard((org) => org.isAdmin)],
     data: {
       titleId: "exportData",
-      organizationPermissions: (org: Organization) => org.isAdmin,
     },
   },
 ];
